Overhaul the server browser.
