+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -566,7 +566,6 @@
     <Compile Include="Widgets\Logic\ConnectionLogic.cs" />
     <Compile Include="Widgets\Logic\FactionTooltipLogic.cs" />
     <Compile Include="Widgets\Logic\CreditsLogic.cs" />
-    <Compile Include="Widgets\Logic\DirectConnectLogic.cs" />
     <Compile Include="Widgets\Logic\DisconnectWatcherLogic.cs" />
     <Compile Include="Widgets\Logic\Ingame\AddFactionSuffixLogic.cs" />
     <Compile Include="Widgets\Logic\Ingame\ClassicProductionLogic.cs" />
@@ -620,7 +619,6 @@
     <Compile Include="Widgets\Logic\MusicPlayerLogic.cs" />
     <Compile Include="Widgets\Logic\PerfDebugLogic.cs" />
     <Compile Include="Widgets\Logic\ReplayBrowserLogic.cs" />
-    <Compile Include="Widgets\Logic\ServerBrowserLogic.cs" />
     <Compile Include="Widgets\Logic\ServerCreationLogic.cs" />
     <Compile Include="Widgets\Logic\SettingsLogic.cs" />
     <Compile Include="Widgets\Logic\SimpleTooltipLogic.cs" />
@@ -706,6 +704,7 @@
     <Compile Include="Widgets\Logic\GlobalChatLogic.cs" />
     <Compile Include="Lint\CheckChromeLogic.cs" />
     <Compile Include="Lint\CheckMapMetadata.cs" />
+    <Compile Include="Widgets\Logic\MultiplayerLogic.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
