--- mods/cnc/chrome/createserver.yaml
@@ -1,135 +1,0 @@
-Container@CREATESERVER_PANEL:
-	Logic: ServerCreationLogic
-	X: (WINDOW_RIGHT - WIDTH)/2
-	Y: (WINDOW_BOTTOM - 260)/2
-	Width: 521
-	Height: 250
-	Children:
-		Label@TITLE:
-			Text: Create Server
-			Width: PARENT_RIGHT
-			Y: 0-25
-			Font: BigBold
-			Contrast: true
-			Align: Center
-		Background@bg:
-			Width: PARENT_RIGHT
-			Height: 215
-			Background: panel-black
-			Children:
-				Background@MAP_BG:
-					X: PARENT_RIGHT-WIDTH-15
-					Y: 15
-					Width: 162
-					Height: 162
-					Background: panel-gray
-					Children:
-						MapPreview@MAP_PREVIEW:
-							X: 1
-							Y: 1
-							Width: 160
-							Height: 160
-				Label@MAP_NAME:
-					X: PARENT_RIGHT-WIDTH-15
-					Y: PARENT_BOTTOM - 33
-					Width: 162
-					Height: 25
-					Align: Center
-					Font: Bold
-				Label@SERVER_NAME_LABEL:
-					X: 15
-					Y: 14
-					Width: 90
-					Height: 25
-					Align: Right
-					Text: Server Name:
-				TextField@SERVER_NAME:
-					X: 110
-					Y: 15
-					Width: 215
-					MaxLength: 50
-					Height: 25
-					Text: My OpenRA Server
-				Label@PASSWORD_LABEL:
-					X: 10
-					Y: 49
-					Width: 95
-					Height: 25
-					Align: Right
-					Text: Password:
-				PasswordField@PASSWORD:
-					X: 110
-					Y: 50
-					Width: 145
-					MaxLength: 20
-					Height: 25
-				Label@AFTER_PASSWORD_LABEL:
-					X: 265
-					Y: 49
-					Width: 95
-					Height: 25
-					Align: Left
-					Text: (optional)
-				Label@LISTEN_PORT_LABEL:
-					X: 15
-					Y: 84
-					Width: 90
-					Height: 25
-					Align: Right
-					Text: Port:
-				TextField@LISTEN_PORT:
-					X: 110
-					Y: 85
-					Width: 50
-					MaxLength: 5
-					Height: 25
-					Text: 1234
-				Label@EXTERNAL_PORT_LABEL:
-					X: 180
-					Y: 84
-					Width: 90
-					Height: 25
-					Align: Right
-					Text: External Port:
-				TextField@EXTERNAL_PORT:
-					X: 275
-					Y: 85
-					Width: 50
-					MaxLength: 5
-					Height: 25
-					Text: 1234
-				Checkbox@ADVERTISE_CHECKBOX:
-					X: 110
-					Y: 135
-					Width: 150
-					Height: 20
-					Font: Regular
-					Text: Advertise Online
-				Checkbox@UPNP_CHECKBOX:
-					X: 110
-					Y: 165
-					Width: 300
-					Height: 20
-					Font: Regular
-					Text: Automatic port forwarding
-		Button@BACK_BUTTON:
-			Key: escape
-			X: 0
-			Y: 214
-			Width: 140
-			Height: 35
-			Text: Back
-		Button@MAP_BUTTON:
-			X: 231
-			Y: 214
-			Width: 140
-			Height: 35
-			Text: Choose Map
-		Button@CREATE_BUTTON:
-			Key: return
-			X: 381
-			Y: 214
-			Width: 140
-			Height: 35
-			Text: Create
-
