--- mods/cnc/chrome/directconnect.yaml
@@ -1,58 +1,0 @@
-Container@DIRECTCONNECT_PANEL:
-	Logic: DirectConnectLogic
-	X: (WINDOW_RIGHT - WIDTH)/2
-	Y: (WINDOW_BOTTOM - 90)/2
-	Width: 370
-	Height: 130
-	Children:
-		Label@TITLE:
-			Width: PARENT_RIGHT
-			Y: 0-25
-			Font: BigBold
-			Contrast: true
-			Align: Center
-			Text: Connect to Server
-		Background@bg:
-			Width: 370
-			Height: 95
-			Background: panel-black
-			Children:
-				Label@ADDRESS_LABEL:
-					X: 50
-					Y: 14
-					Width: 95
-					Height: 25
-					Align: Right
-					Text: Address:
-				TextField@IP:
-					X: 150
-					Y: 15
-					Width: 200
-					Height: 25
-				Label@PORT_LABEL:
-					X: 50
-					Y: 49
-					Width: 95
-					Height: 25
-					Align: Right
-					Text: Port:
-				TextField@PORT:
-					X: 150
-					Y: 50
-					Width: 200
-					Height: 25
-		Button@BACK_BUTTON:
-			Key: escape
-			X: 0
-			Y: 94
-			Width: 140
-			Height: 35
-			Text: Back
-		Button@JOIN_BUTTON:
-			Key: return
-			X: 230
-			Y: 94
-			Width: 140
-			Height: 35
-			Text: Join
-
