--- mods/cnc/chrome/serverbrowser.yaml
@@ -1,170 +1,0 @@
-Container@SERVERBROWSER_PANEL:
-	Logic: ServerBrowserLogic
-	X: (WINDOW_RIGHT - WIDTH)/2
-	Y: (WINDOW_BOTTOM - HEIGHT)/2
-	Width: 730
-	Height: 595
-	Children:
-		Label@TITLE:
-			Text: Multiplayer
-			Width: 740
-			Y: 0-10
-			Font: BigBold
-			Contrast: true
-			Align: Center
-		Background@bg:
-			Width: 730
-			Height: PARENT_BOTTOM - 30
-			Background: panel-black
-			Y: 15
-			Children:
-				Container@GLOBALCHAT_ROOT:
-					X: 15
-					Y: 15
-					Width: 700
-					Height: 260
-				ScrollPanel@SERVER_LIST:
-					X: 15
-					Y: 280
-					Width: 700
-					Height: 240
-					Children:
-						ScrollItem@HEADER_TEMPLATE:
-							Width: PARENT_RIGHT-27
-							Height: 25
-							X: 2
-							Visible: false
-							Children:
-								Label@LABEL:
-									Y: 0-1
-									Font: Bold
-									Width: PARENT_RIGHT
-									Height: 25
-									Align: Center
-						ScrollItem@SERVER_TEMPLATE:
-							Width: PARENT_RIGHT-27
-							Height: 68
-							X: 2
-							Y: 0
-							Children:
-								MapPreview@MAP_PREVIEW:
-									X: 2
-									Y: 2
-									Width: 64
-									Height: 64
-									ShowSpawnPoints: no
-								Label@TITLE:
-									X: 70
-									Width: 200
-									Height: 25
-									Font: Bold
-								Label@MAP:
-									X: 70
-									Y: 20
-									Width: 250
-									Height: 25
-								Label@PLAYERS:
-									X: 70
-									Y: 40
-									Width: 50
-									Height: 25
-								Label@STATE:
-									Width: 140
-									X: PARENT_RIGHT-150
-									Align: Right
-									Height: 25
-									Font: Bold
-								Label@IP:
-									Width: 140
-									X: PARENT_RIGHT-150
-									Y: 20
-									Align: Right
-									Height: 25
-								Label@LOCATION:
-									Width: 140
-									X: PARENT_RIGHT-150
-									Y: 40
-									Align: Right
-									Height: 25
-				Label@PROGRESS_LABEL:
-					X: (PARENT_RIGHT - WIDTH) / 2
-					Y: PARENT_BOTTOM / 2 - HEIGHT + (280 / 2)
-					Width: 710
-					Height: 25
-					Font: Bold
-					Align: Center
-					Visible: false
-				Label@SHOW_LABEL_TITLE:
-					X: 20
-					Y: 525
-					Width: 20
-					Height: 25
-					Text: Show:
-					Font: Bold
-				Checkbox@WAITING_FOR_PLAYERS:
-					X: 80
-					Y: 527
-					Width: 100
-					Height: 20
-					Text: Waiting
-					TextColor: 50,205,50
-				Checkbox@EMPTY:
-					X: 180
-					Y: 527
-					Width: 100
-					Height: 20
-					Text: Empty
-				Checkbox@PASSWORD_PROTECTED:
-					X: 270
-					Y: 527
-					Width: 100
-					Height: 20
-					Text: Protected
-					TextColor: 255,0,0
-				Checkbox@ALREADY_STARTED:
-					X: 385
-					Y: 527
-					Width: 100
-					Height: 20
-					Text: Started
-					TextColor: 255,165,0
-				Checkbox@INCOMPATIBLE_VERSION:
-					X: 480
-					Y: 527
-					Width: 100
-					Height: 20
-					Text: Incompatible
-					TextColor: 190,190,190
-				Button@REFRESH_BUTTON:
-					X: PARENT_RIGHT - WIDTH - 15
-					Y: 525
-					Width: 100
-					Height: 25
-					Text: Refresh
-		Button@BACK_BUTTON:
-			Key: escape
-			X: 0
-			Y: PARENT_BOTTOM - 16
-			Width: 140
-			Height: 35
-			Text: Back
-		Button@CREATE_BUTTON:
-			X: PARENT_RIGHT - 140 - 10 - 140 - 10 - 140
-			Y: PARENT_BOTTOM - 16
-			Width: 140
-			Height: 35
-			Text: Create
-		Button@DIRECTCONNECT_BUTTON:
-			X: PARENT_RIGHT - 140 - 10 - 140
-			Y: PARENT_BOTTOM - 16
-			Width: 140
-			Height: 35
-			Text: Direct IP
-		Button@JOIN_BUTTON:
-			Key: return
-			X: PARENT_RIGHT - 140
-			Y: PARENT_BOTTOM - 16
-			Width: 140
-			Height: 35
-			Text: Join
-
