+++ mods/d2k/mod.yaml
@@ -93,11 +93,12 @@ ChromeLayout:
 	./mods/ra/chrome/lobby-globalchat.yaml
 	./mods/d2k/chrome/color-picker.yaml
 	./mods/ra/chrome/map-chooser.yaml
-	./mods/ra/chrome/create-server.yaml
-	./mods/ra/chrome/serverbrowser.yaml
+	./mods/ra/chrome/multiplayer.yaml
+	./mods/ra/chrome/multiplayer-browser.yaml
+	./mods/ra/chrome/multiplayer-createserver.yaml
+	./mods/ra/chrome/multiplayer-directconnect.yaml
+	./mods/ra/chrome/multiplayer-globalchat.yaml
 	./mods/ra/chrome/connection.yaml
-	./mods/ra/chrome/directconnect.yaml
-	./mods/ra/chrome/replaybrowser.yaml
 	./mods/d2k/chrome/dropdowns.yaml
 	./mods/ra/chrome/musicplayer.yaml
 	./mods/d2k/chrome/tooltips.yaml
@@ -105,7 +106,6 @@ ChromeLayout:
 	./mods/d2k/chrome/missionbrowser.yaml
 	./mods/ra/chrome/confirmation-dialogs.yaml
 	./mods/ra/chrome/editor.yaml
-	./mods/ra/chrome/globalchat.yaml
 
 Weapons:
 	./mods/d2k/weapons.yaml
