--- mods/ra/chrome/serverbrowser.yaml
@@ -1,170 +1,0 @@
-Background@SERVERBROWSER_PANEL:
-	Logic: ServerBrowserLogic
-	X: (WINDOW_RIGHT - WIDTH)/2
-	Y: (WINDOW_BOTTOM - HEIGHT)/2
-	Width: 740
-	Height: 645
-	Children:
-		Label@MULTIPLAYER_LABEL_TITLE:
-			X: 0
-			Y: 15
-			Width: PARENT_RIGHT
-			Height: 25
-			Text: Multiplayer
-			Align: Center
-			Font: Bold
-		Label@SHOW_LABEL_TITLE:
-			X: 20
-			Y: 48
-			Width: 20
-			Height: 25
-			Text: Show:
-			Font: Bold
-		Checkbox@WAITING_FOR_PLAYERS:
-			X: 80
-			Y: 50
-			Width: 100
-			Height: 20
-			Text: Waiting
-			TextColor: 50,205,50
-		Checkbox@EMPTY:
-			X: 180
-			Y: 50
-			Width: 100
-			Height: 20
-			Text: Empty
-		Checkbox@PASSWORD_PROTECTED:
-			X: 270
-			Y: 50
-			Width: 100
-			Height: 20
-			Text: Protected
-			TextColor: 255,0,0
-		Checkbox@ALREADY_STARTED:
-			X: 385
-			Y: 50
-			Width: 100
-			Height: 20
-			Text: Started
-			TextColor: 255,165,0
-		Checkbox@INCOMPATIBLE_VERSION:
-			X: 480
-			Y: 50
-			Width: 100
-			Height: 20
-			Text: Incompatible
-			TextColor: 190,190,190
-		ScrollPanel@SERVER_LIST:
-			X: 20
-			Y: 80
-			Width: 700
-			Height: 240
-			Children:
-				ScrollItem@HEADER_TEMPLATE:
-					BaseName: scrollheader
-					Width: PARENT_RIGHT-27
-					Height: 25
-					X: 2
-					Visible: false
-					Children:
-						Label@LABEL:
-							Y: 0-1
-							Font: Bold
-							Width: PARENT_RIGHT
-							Height: 25
-							Align: Center
-				ScrollItem@SERVER_TEMPLATE:
-					Width: PARENT_RIGHT-27
-					Height: 68
-					X: 2
-					Y: 0
-					Children:
-						MapPreview@MAP_PREVIEW:
-							X: 2
-							Y: 2
-							Width: 64
-							Height: 64
-							ShowSpawnPoints: no
-						Label@TITLE:
-							X: 70
-							Width: 200
-							Height: 25
-							Font: Bold
-						Label@MAP:
-							X: 70
-							Y: 20
-							Width: 250
-							Height: 25
-						Label@PLAYERS:
-							X: 70
-							Y: 40
-							Width: 50
-							Height: 25
-						Label@STATE:
-							Width: 140
-							X: PARENT_RIGHT-150
-							Align: Right
-							Height: 25
-							Font: Bold
-						Label@IP:
-							Width: 140
-							X: PARENT_RIGHT-150
-							Y: 20
-							Align: Right
-							Height: 25
-						Label@LOCATION:
-							Width: 140
-							X: PARENT_RIGHT-150
-							Y: 40
-							Align: Right
-							Height: 25
-		Container@GLOBALCHAT_ROOT:
-			X: 20
-			Y: 370
-			Width: 700
-			Height: 255
-		Label@PROGRESS_LABEL:
-			X: (PARENT_RIGHT - WIDTH) / 2
-			Y: PARENT_BOTTOM / 2 - HEIGHT
-			Width: 150
-			Height: 30
-			Text: Fetching games...
-			Align: Center
-		Button@REFRESH_BUTTON:
-			X: 20
-			Y: 325
-			Width: 100
-			Height: 25
-			Text: Refresh
-			Font: Bold
-		Button@CREATE_BUTTON:
-			X: PARENT_RIGHT - 120 - 120 - 120 - 120
-			Y: 325
-			Width: 100
-			Height: 25
-			Text: Create
-			Font: Bold
-		Button@DIRECTCONNECT_BUTTON:
-			X: PARENT_RIGHT - 120 - 120 - 120
-			Y: 325
-			Width: 100
-			Height: 25
-			Text: Direct IP
-			Font: Bold
-		Button@JOIN_BUTTON:
-			X: PARENT_RIGHT - 120 - 120
-			Y: 325
-			Width: 100
-			Height: 25
-			Text: Join
-			Font: Bold
-			Key: return
-		Button@BACK_BUTTON:
-			X: PARENT_RIGHT - 120
-			Y: 325
-			Width: 100
-			Height: 25
-			Text: Cancel
-			Font: Bold
-			Key: escape
-
