+++ mods/ra/mod.yaml
@@ -104,10 +104,12 @@ ChromeLayout:
 	./mods/ra/chrome/lobby-globalchat.yaml
 	./mods/ra/chrome/color-picker.yaml
 	./mods/ra/chrome/map-chooser.yaml
-	./mods/ra/chrome/create-server.yaml
-	./mods/ra/chrome/serverbrowser.yaml
+	./mods/ra/chrome/multiplayer.yaml
+	./mods/ra/chrome/multiplayer-browser.yaml
+	./mods/ra/chrome/multiplayer-createserver.yaml
+	./mods/ra/chrome/multiplayer-directconnect.yaml
+	./mods/ra/chrome/multiplayer-globalchat.yaml
 	./mods/ra/chrome/connection.yaml
-	./mods/ra/chrome/directconnect.yaml
 	./mods/ra/chrome/replaybrowser.yaml
 	./mods/ra/chrome/dropdowns.yaml
 	./mods/ra/chrome/musicplayer.yaml
@@ -116,7 +118,6 @@ ChromeLayout:
 	./mods/ra/chrome/missionbrowser.yaml
 	./mods/ra/chrome/confirmation-dialogs.yaml
 	./mods/ra/chrome/editor.yaml
-	./mods/ra/chrome/globalchat.yaml
 
 Weapons:
 	./mods/ra/weapons/explosions.yaml
